refactor(agent): split oversized service modules
